# Fix bugs in penalty scores evaluation