--- a/lib/core/mask-pattern.js
+++ b/lib/core/mask-pattern.js
@@ -17,7 +17,7 @@ exports.Patterns = {
  * Weighted penalty scores for the undesirable features
  * @type {Object}
  */
-var PenalityScores = {
+var PenaltyScores = {
   N1: 3,
   N2: 3,
   N3: 40,
@@ -25,62 +25,72 @@ var PenalityScores = {
 }
 
 /**
- * Find adjacent modules in row/column with the same color
- * and assign a penality value.
- *
- * Points: N1 + i
- * i is the amount by which the number of adjacent modules of the same color exceeds 5
- */
-function getPenalityN1 (data) {
+* Find adjacent modules in row/column with the same color
+* and assign a penalty value.
+*
+* Points: N1 + i
+* i is the amount by which the number of adjacent modules of the same color exceeds 5
+*/
+exports.getPenaltyN1 = function getPenaltyN1 (data) {
   var size = data.size
   var points = 0
+  var sameCountCol = 0
+  var sameCountRow = 0
+  var lastCol = null
+  var lastRow = null
 
   for (var row = 0; row < size; row++) {
+    sameCountCol = sameCountRow = 0
+    lastCol = lastRow = null
+
     for (var col = 0; col < size; col++) {
-      // number of consecutive modules with same color
-      var sameCount = 0
-      var dark = data.get(row, col)
-
-      for (var r = -1; r <= 1; r++) {
-        if (row + r < 0 || size <= row + r) continue
-
-        for (var c = -1; c <= 1; c++) {
-          if (col + c < 0 || size <= col + c) continue
-          if (r === 0 && c === 0) continue
-          if (dark === data.get(row + r, col + c)) sameCount++
-        }
+      var module = data.get(row, col)
+      if (module === lastCol) {
+        sameCountCol++
+      } else {
+        if (sameCountCol >= 5) points += PenaltyScores.N1 + (sameCountCol - 5)
+        lastCol = module
+        sameCountCol = 1
       }
 
-      if (sameCount > 5) {
-        points += PenalityScores.N1 + sameCount - 5
+      module = data.get(col, row)
+      if (module === lastRow) {
+        sameCountRow++
+      } else {
+        if (sameCountRow >= 5) points += PenaltyScores.N1 + (sameCountRow - 5)
+        lastRow = module
+        sameCountRow = 1
       }
     }
+
+    if (sameCountCol >= 5) points += PenaltyScores.N1 + (sameCountCol - 5)
+    if (sameCountRow >= 5) points += PenaltyScores.N1 + (sameCountRow - 5)
   }
 
   return points
 }
 
 /**
- * Find 2x2 blocks with the same color and assign a penality value
+ * Find 2x2 blocks with the same color and assign a penalty value
  *
  * Points: N2 * (m - 1) * (n - 1)
  */
-function getPenalityN2 (data) {
+exports.getPenaltyN2 = function getPenaltyN2 (data) {
   var size = data.size
   var points = 0
 
   for (var row = 0; row < size - 1; row++) {
     for (var col = 0; col < size - 1; col++) {
-      var count = 0
-      if (data.get(row, col)) count++
-      if (data.get(row + 1, col)) count++
-      if (data.get(row, col + 1)) count++
-      if (data.get(row + 1, col + 1)) count++
-      if (count === 0 || count === 4) points += PenalityScores.N2
+      var last = data.get(row, col) +
+        data.get(row, col + 1) +
+        data.get(row + 1, col) +
+        data.get(row + 1, col + 1)
+
+      if (last === 4 || last === 0) points++
     }
   }
 
-  return points
+  return points * PenaltyScores.N2
 }
 
 /**
@@ -89,40 +99,24 @@ function getPenalityN2 (data) {
  *
  * Points: N3 * number of pattern found
  */
-function getPenalityN3 (data) {
+exports.getPenaltyN3 = function getPenaltyN3 (data) {
   var size = data.size
   var points = 0
-  var row, col
+  var bitsCol = 0
+  var bitsRow = 0
 
-  for (row = 0; row < size; row++) {
-    for (col = 0; col < size - 6; col++) {
-      if (data.get(row, col) &&
-         !data.get(row, col + 1) &&
-          data.get(row, col + 2) &&
-          data.get(row, col + 3) &&
-          data.get(row, col + 4) &&
-         !data.get(row, col + 5) &&
-          data.get(row, col + 6)) {
-        points += PenalityScores.N3
-      }
+  for (var row = 0; row < size; row++) {
+    bitsCol = bitsRow = 0
+    for (var col = 0; col < size; col++) {
+      bitsCol = ((bitsCol << 1) & 0x7FF) | data.get(row, col)
+      if (col >= 10 && (bitsCol === 0x5D0 || bitsCol === 0x05D)) points++
+
+      bitsRow = ((bitsRow << 1) & 0x7FF) | data.get(col, row)
+      if (col >= 10 && (bitsRow === 0x5D0 || bitsRow === 0x05D)) points++
     }
   }
 
-  for (col = 0; col < size; col++) {
-    for (row = 0; row < size - 6; row++) {
-      if (data.get(row, col) &&
-         !data.get(row + 1, col) &&
-          data.get(row + 2, col) &&
-          data.get(row + 3, col) &&
-          data.get(row + 4, col) &&
-         !data.get(row + 5, col) &&
-          data.get(row + 6, col)) {
-        points += PenalityScores.N3
-      }
-    }
-  }
-
-  return points
+  return points * PenaltyScores.N3
 }
 
 /**
@@ -133,18 +127,15 @@ function getPenalityN3 (data) {
  * k is the rating of the deviation of the proportion of dark modules
  * in the symbol from 50% in steps of 5%
  */
-function getPenalityN4 (data) {
+exports.getPenaltyN4 = function getPenaltyN4 (data) {
   var darkCount = 0
-  var size = data.size
+  var modulesCount = data.data.length
 
-  for (var col = 0; col < size; col++) {
-    for (var row = 0; row < size; row++) {
-      if (data.get(row, col)) darkCount++
-    }
-  }
+  for (var i = 0; i < modulesCount; i++) darkCount += data.data[i]
 
-  var ratio = Math.abs(100 * darkCount / size / size - 50) / 5
-  return ratio * PenalityScores.N4
+  var k = Math.abs(Math.ceil((darkCount * 100 / modulesCount) / 5) - 10)
+
+  return k * PenaltyScores.N4
 }
 
 /**
@@ -193,26 +184,27 @@ exports.applyMask = function applyMask (pattern, data) {
  * @param  {BitMatrix} data
  * @return {Number} Mask pattern reference number
  */
-exports.getBestMask = function getBestMask (data) {
+exports.getBestMask = function getBestMask (data, setupFormatFunc) {
   var numPatterns = Object.keys(exports.Patterns).length
   var bestPattern = 0
-  var lowerPenality = Infinity
+  var lowerPenalty = Infinity
 
   for (var p = 0; p < numPatterns; p++) {
+    setupFormatFunc(p)
     exports.applyMask(p, data)
 
-    // Calculate penality
-    var penality =
-      getPenalityN1(data) +
-      getPenalityN2(data) +
-      getPenalityN3(data) +
-      getPenalityN4(data)
+    // Calculate penalty
+    var penalty =
+      exports.getPenaltyN1(data) +
+      exports.getPenaltyN2(data) +
+      exports.getPenaltyN3(data) +
+      exports.getPenaltyN4(data)
 
     // Undo previously applied mask
     exports.applyMask(p, data)
 
-    if (penality < lowerPenality) {
-      lowerPenality = penality
+    if (penalty < lowerPenalty) {
+      lowerPenalty = penalty
       bestPattern = p
     }
   }