--- a/lib/core/qrcode.js
+++ b/lib/core/qrcode.js
@@ -122,9 +122,8 @@ function setupAlignmentPattern (matrix, version) {
  *
  * @param  {BitMatrix} matrix  Modules matrix
  * @param  {Number}    version QR Code version
- * @param  {Boolean}   reserve If true, marks bits as reserved and set their values to 0
  */
-function setupVersionInfo (matrix, version, reserve) {
+function setupVersionInfo (matrix, version) {
   var size = matrix.size
   var bits = Version.getEncodedBits(version)
   var row, col, mod
@@ -132,7 +131,7 @@ function setupVersionInfo (matrix, version, reserve) {
   for (var i = 0; i < 18; i++) {
     row = Math.floor(i / 3)
     col = i % 3 + size - 8 - 3
-    mod = (!reserve && ((bits >> i) & 1) === 1)
+    mod = ((bits >> i) & 1) === 1
 
     matrix.set(row, col, mod, true)
     matrix.set(col, row, mod, true)
@@ -145,15 +144,14 @@ function setupVersionInfo (matrix, version, reserve) {
  * @param  {BitMatrix} matrix               Modules matrix
  * @param  {ErrorCorrectionLevel}    errorCorrectionLevel Error correction level
  * @param  {Number}    maskPattern          Mask pattern reference value
- * @param  {Boolean}   reserve              If true, marks bits as reserved and set their values to 0
  */
-function setupFormatInfo (matrix, errorCorrectionLevel, maskPattern, reserve) {
+function setupFormatInfo (matrix, errorCorrectionLevel, maskPattern) {
   var size = matrix.size
   var bits = FormatInfo.getEncodedBits(errorCorrectionLevel, maskPattern)
   var i, mod
 
   for (i = 0; i < 15; i++) {
-    mod = (!reserve && ((bits >> i) & 1) === 1)
+    mod = ((bits >> i) & 1) === 1
 
     // vertical
     if (i < 6) {
@@ -175,7 +173,7 @@ function setupFormatInfo (matrix, errorCorrectionLevel, maskPattern, reserve) {
   }
 
   // fixed module
-  matrix.set(size - 8, 8, !reserve, true)
+  matrix.set(size - 8, 8, 1, true)
 }
 
 /**
@@ -433,31 +431,28 @@ function createSymbol (data, version, errorCorrectionLevel) {
   setupTimingPattern(modules)
   setupAlignmentPattern(modules, version)
 
-  // Add temporary blank bits for format info and version info just to set them as reserved.
+  // Add temporary dummy bits for format info just to set them as reserved.
   // This is needed to prevent these bits from being masked by {@link MaskPattern.applyMask}
   // since the masking operation must be performed only on the encoding region.
   // These blocks will be replaced with correct values later in code.
-  setupFormatInfo(modules, errorCorrectionLevel, 0, true)
+  setupFormatInfo(modules, errorCorrectionLevel, 0)
+
   if (version >= 7) {
-    setupVersionInfo(modules, version, true)
+    setupVersionInfo(modules, version)
   }
 
   // Add data codewords
   setupData(modules, dataBits)
 
   // Find best mask pattern
-  var maskPattern = MaskPattern.getBestMask(modules)
+  var maskPattern = MaskPattern.getBestMask(modules,
+    setupFormatInfo.bind(null, modules, errorCorrectionLevel))
 
   // Apply mask pattern
   MaskPattern.applyMask(maskPattern, modules)
 
   // Replace format info bits with correct values
   setupFormatInfo(modules, errorCorrectionLevel, maskPattern)
-
-  // Replace version info bits with correct values
-  if (version >= 7) {
-    setupVersionInfo(modules, version)
-  }
 
   return {
     modules: modules,