… test/e2e/toDataURL.test.js
… test/e2e/toFile.test.js
… test/e2e/toString.test.js