I've noticed that `getBestMask()` function often returns a wrong pattern.

Comparing the qrcode generated by this lib with a qrcode returned by [googleapis](https://chart.googleapis.com/chart?chs=300x300&cht=qr&chld=L&chl=text) (which I took as reference) it will look quite different.

Digging a bit I found that penalty scores are calculated in a wrong way and as a consequence a wrong mask pattern is applied to the symbol.

In details:
- the current implementation of `getPenaltyN1()` and `getPenaltyN2()` doesn't take into account light modules but only the dark ones.
- `getPenaltyN3()` doesn't look for the correct pattern but only part of it.
- `getPenaltyN4()` doesn't calculate the correct ratio of light modules to dark modules.
- Penalty scores are evaluated without taking into account format and version info.

This PR fixes these functions and adds a few tests for them.

Since best mask calculation is where most of the time is spent during symbol creation, I also tried to improve performances a bit.

I ran a quick benchmark test to see if something was actually improved.
Here are the results (on a very old Intel Core Duo laptop):

| Function    | Before (ops/sec) | After (ops/sec) |
|-------------|:----------------:|:---------------:|
|getPenaltyN1 | 16,583           | 49,448          |
|getPenaltyN2 | 119,755          | 183,098         |
|getPenaltyN3 | 109,634          | 204,088         |
|getPenaltyN4 | 197,187          | 589,327         |

And qr code creation:

| Function     | Before (ops/sec) | After (ops/sec) |
|--------------|:----------------:|:---------------:|
|QRCode.create | 861              | 1,319           |

Not a huge improvement, but it's still something.